WIP
Getting closer, pretty sure my math is off in some places when
calculating array indices, need to go back and verify that code.

We're also going to have a hard time mapping the array indices
in the filterFrom() method when the user is requesting more
entries than there actually are.

So in other words, this code is really buggy...