Passing tests for all filters
Ended up rewriting the way that we handle journal entries, both
in-memory and on-disk. Instead of having a list of journal entries
and doing a binary search to find the correct entries, we now have
what is basically a B-Tree that breaks entries down by year, month,
and day.

That required rewriting all the filters, which then made FilterResult
useless since we no longer need to keep track of position in a list,
which then required updates to tests and a couple of other places
since filters now just return the entries directly.

Things are hopelessly broken right now, there's no saving to disk yet
since I need to get that in place again. I'm also not entirely
convinced that the filters are all correct, even if the tests are all
passing now. We'll need to test with data across multiple months and
years to be absolutely sure that what we have works.