Support for filtering journal entries that fall on a specific date (#11)
Ended up rewriting the way that we handle journal entries, both
in-memory and on-disk. Instead of having a list of journal entries
and doing a binary search to find the correct entries, we now have
what is basically a B-Tree that breaks entries down by year, month,
and day.

That required rewriting all the filters, which then made FilterResult
useless since we no longer need to keep track of the position in a list,
which then required updates to tests and a couple of other places
since filters now just return the entries directly.