Editing journal entries causes dup date line
This fixes #12. In a nutshell, we were always creating a temp
file when adding or editing a journal by concatenating the journal
entry date with the body. This meant we were doubling, tripling,
quadrupling, etc. the journal entry date on the first line each
time we edited the file.

Simple fix, just make the code for writing the journal entry to the
temp file for editing aware of whether a new journal entry or an
existing one is being edited.