Editing journal entries causes dup date line (#14)
* Editing journal entries causes dup date line

This fixes #12. In a nutshell, we were always creating a temp
file when adding or editing a journal by concatenating the journal
entry date with the body. This meant we were doubling, tripling,
quadrupling, etc. the journal entry date on the first line each
time we edited the file.

Simple fix, just make the code for writing the journal entry to the
temp file for editing aware of whether a new journal entry or an
existing one is being edited.

* Refactoring temp file logic into new class

This is to make testing a little easier, need to refactor
EditorService more for testability but that will have to be a task
for later

* Removing tempfile-parser.ts

This got merged into the source file for the new
JournalTempFile class

* Removing unused import statements