Need way to configure the editor used for journal entries
Fixes #13

Probably not the best code I've written but it appears to work. In a
nutshell, we add a new ConfigService with a couple of different
methods:

- getStoragePath() method for returning where the app stores files
- getOrCreateConfigFile() method for returning the current config
  file parsed from disk, or creating a new config file with some
  sane defaults when one doesn't exist

We then update the `Application` class to use `ConfigService` to get
the storage location instead of determining it one its own.

We also update `EditorService` to use `ConfigService` as well to
determine which command to execute for the editor instead of determining
this on its own.

Other than that, we just need to wire up the `ConfigService` class with
the IoC container so it can get constructor injected where needed.