Improve date format used for journal entries
Fixes #1

Updated the program to store the date inside the journal entries
as "MM-DD-YYYY hh:ii A" but to internally store the date in our
data structures using ISO 8601 for backwards compatibility.

Any existing journal entries will continue to show up with the old
ISO 8601 date format but new journal entries will be created using the
new format.

We should move away from storing the journal date as part of the
journal body, instead inserting it into the body when we show a
journal entry or edit an existing journal entry.